ProcessInput added instead of Move
To make the state machine easier, I moved the controls from the player
move into the Game1 class and added a processInput to contain
everything. I also added some of the enum states in it. Some of it might
be buggy. Once we have everything it will be easier to test it and see
if it will work.